Move passthrough recorder into vcr
